+++ src/main/java/to/kit/shooter/entity/VisualView.java
@@ -26,7 +26,7 @@ public class VisualView {
 	private String owner;
 	private int access;
 	private int visualType;
-	private int visualSeq;
+	private long visualSeq;
 	private String orientation;
 	private String name;
 	private int imagelen;
